port check 056: check_font_has_latest_ttfautohint_applied
(issue #1388)
