+++ Lib/fontbakery/checks.py
@@ -250,78 +250,6 @@ def perform_all_fontforge_checks(fb, validation_state):
            "Hinds do not overlap.")
 
 
-def check_font_has_latest_ttfautohint_applied(fb, font, ttfautohint_missing):
-  fb.new_check("056", "Font has old ttfautohint applied?")
-  ''' ----------------------------------------------------
-     Font has old ttfautohint applied ?
-
-     1. find which version was used, grepping the name table or reading
-        the ttfa table (which are created if the `-I` or `-t` args
-        respectively were passed to ttfautohint, to record its args in
-        the ttf file) (there is a pypi package
-        https://pypi.python.org/pypi/font-ttfa for reading the ttfa table,
-        although per https://github.com/source-foundry/font-ttfa/issues/1
-        it might be better to inline the code... :)
-
-     2. find which version of ttfautohint is installed
-        and warn if not available, similar to ots check above
-
-     3. rehint the font with the latest version of ttfautohint
-        using the same options
-  '''
-
-  def ttfautohint_version(values):
-    for value in values:
-      results = re.search(r'ttfautohint \(v(.*)\)', value)
-      if results:
-        return results.group(1)
-
-  def installed_ttfa_version(value):
-    return re.search(r'ttfautohint ([^-\n]*)(-.*)?\n', value).group(1)
-
-  def installed_version_is_newer(installed, used):
-    installed = map(int, installed.split("."))
-    used = map(int, used.split("."))
-    return installed > used
-
-  version_strings = get_name_string(font, NAMEID_VERSION_STRING)
-  ttfa_version = ttfautohint_version(version_strings)
-  if len(version_strings) == 0:
-    fb.error("This font file lacks mandatory "
-             "version strings in its name table.")
-  elif ttfa_version is None:
-    fb.info(("Could not detect which version of"
-             " ttfautohint was used in this font."
-             " It is typically specified as a comment"
-             " in the font version entry of the 'name' table."
-             " Font version string is: '{}'").format(version_strings[0]))
-  elif ttfautohint_missing:
-    fb.skip("This check requires ttfautohint"
-            " to be available in the system.")
-  else:
-    import subprocess
-    ttfa_cmd = ["ttfautohint",
-                "-V"]  # print version info
-    ttfa_output = subprocess.check_output(ttfa_cmd,
-                                          stderr=subprocess.STDOUT)
-    installed_ttfa = installed_ttfa_version(ttfa_output)
-    try:
-      if installed_version_is_newer(installed_ttfa,
-                                    ttfa_version):
-        fb.info(("ttfautohint used in font = {};"
-                 " installed = {}; Need to re-run"
-                 " with the newer version!").format(ttfa_version,
-                                                    installed_ttfa))
-      else:
-        fb.ok("ttfautohint available in the system is older"
-              " than the one used in the font.")
-    except:
-      fb.error(("failed to parse ttfautohint version strings:\n"
-                "  * installed = '{}'\n"
-                "  * used = '{}'").format(installed_ttfa,
-                                          ttfa_version))
-
-
 def check_name_table_entries_do_not_contain_linebreaks(fb, font):
   fb.new_check("057", "Name table entries should not contain line-breaks")
   failed = False
